pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TodoList"
version = "0.1.0"
description = "TodoList 만들기"
requires-python = ">=3.10"

dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "bcrypt==3.2.2",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "click==8.3.0",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.120.0",
    "fastapi-cli==0.0.14",
    "fastapi-cloud-cli==0.3.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "orjson==3.11.4",
    "passlib==1.7.4",
    "pyasn1==0.6.1",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-extra-types==2.10.6",
    "pydantic-settings==2.11.0",
    "pydantic_core==2.41.4",
    "PyMySQL==1.1.2",
    "python-dotenv==1.2.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "PyYAML==6.0.3",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.44",
    "starlette==0.48.0",
    "typer==0.20.0",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "ujson==5.11.0",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "watchfiles==1.1.1",
    "websockets==15.0.1"
]

